requirements.txt
discord.py==2.3.2
python-dotenv==1.0.1
aiohttp>=3.9.0
PyNaCl>=1.5.0
cffi>=1.15.0